requirements.txt
arrow==1.2.3
asgiref==3.7.2
binaryornot==0.4.4
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cookiecutter==2.3.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
djangorestframework==3.14.0
Flask==2.3.2
flast==1.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lib==4.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysql==0.0.3
mysqlclient==2.2.0
oauthlib==3.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.7
